backend/requirements.txt
Flask
uwsgi
line-bot-sdk
opencv-python
Pillow
psycopg2